added: market area, lots of comments, customcomponent class for certain controllers for dynamic upgrade system, dynamic upgrade system, buying of health

+++ Bridg3D/Assets/Scripts/Coin.cs
@@ -4,10 +4,11 @@ using UnityEngine;
 
 public class Coin : MonoBehaviour
 {
-    public int value = 1;
+    public float value = 1;
 
     void OnTriggerEnter(Collider other)
     {
+        //only go to player
         WalletController walletController = other.GetComponent<WalletController>();
         if(!walletController)
             return;
